feat: implemented Facade pattern to LaunchCore class

+++ Configuration/AppConfig.cs
@@ -3,7 +3,7 @@ namespace GameLibrary.Configuration
   public class AppConfig
   {
     public required string DownloadPath { get; set; }
-    public int MaxDownloadSpeed { get; set; }
+    public double MaxDownloadSpeed { get; set; }
     public bool EnableAutoUpdates { get; set; }
     public Theme CurrentTheme { get; set; }
